% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_det_menv.R
\name{alg_det_menv}
\alias{alg_det_menv}
\title{Template for new functions}
\usage{
alg_det_menv(
  rap,
  max_T = 1e-05,
  verbose = 2,
  debug = FALSE,
  debug_trinity = FALSE,
  random_trinity_selection = FALSE
)
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
This is a template.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{Warning}{

This is a warning
}

\examples{
Some examples
}
