% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_det_pdp.R
\name{alg_det_pdp}
\alias{alg_det_pdp}
\title{Template for new functions}
\usage{
alg_det_pdp(rap_environment, max_T = 100)
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
This is a template.
}
\section{Warning}{

This is a warning
}

\examples{
Some examples
}
