% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_gillespie_kernel.R
\name{alg_gillespie_kernel}
\alias{alg_gillespie_kernel}
\title{Template for new functions}
\usage{
alg_gillespie_kernel(rules)
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
This is a template.
}
\section{Warning}{

This is a warning
}

\examples{
Some examples
}
