% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_gillepies_menv.R
\name{alg_gillespie_menv}
\alias{alg_gillespie_menv}
\title{Template for new functions}
\usage{
alg_gillespie_menv(rap_environment, max_T = 10)
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
This is a template.
}
\section{Warning}{

This is a warning
}

\examples{
Some examples
}
