% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rule.R
\name{apply_rule}
\alias{apply_rule}
\title{Template for new functions}
\usage{
apply_rule(my_param)
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
This is a template.
}
\section{Warning}{

This is a warning
}

\examples{
Some examples
}
