% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rule_menv.R
\name{apply_rule_menv}
\alias{apply_rule_menv}
\title{Apply a rule to a given multienvironmental P system}
\usage{
apply_rule_menv(rap, rule_id, comp_id, verbose = FALSE, debug = FALSE)
}
\arguments{
\item{rap}{A rap object.}

\item{rule_id}{The id of the rule to be applied.}

\item{comp_id}{The id of the affected compartment.}

\item{verbose}{The verbosity, between 0 and 5.}

\item{debug}{Useful if you want to debug the execution of the function.}
}
\value{
A new rap object, the result of applying the given rule.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Apply a rule to a given multienvironmental P system given as a \code{rap} object.
}
\section{TODO}{

\itemize{
\item Print the trace of the execution, perhaps with the \code{RAPS::show_rap()} function.
\item Include examples
}
}

\seealso{
\code{apply_rules} to apply more than one rule at the same time, \code{apply_rule} to apply one rule in a monoenvironmental P system.
}
