% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rule_pdp.R
\name{apply_rule_pdp}
\alias{apply_rule_pdp}
\title{Apply a rule to a given rap object}
\usage{
apply_rule_pdp(rap, rule_id, membrane_id)
}
\arguments{
\item{rap}{A rap object-}

\item{rule_id}{The id of the rule to be applied-}
}
\value{
A new rap object, the result of applying the given rule-
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
TODO
}
\seealso{
\code{apply_rules} to apply more than one rule at the same time.
#' @section TODO:
\itemize{
\item Print the trace of the execution, perhaps with the \code{RAPS::show_rap()} function.
}
}
