% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rules.R
\name{apply_rules}
\alias{apply_rules}
\title{Template for new functions}
\usage{
apply_rules(rap, rules, return_between = FALSE)
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
This is a template.
}
\section{Warning}{

This is a warning
}

\examples{
Some examples
}
