% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rules_sequentially.R
\name{apply_rules_sequentially}
\alias{apply_rules_sequentially}
\title{Apply one or more rules to a given P system}
\usage{
apply_rules_sequentially(rap, rules_ids, return_between = FALSE)
}
\arguments{
\item{rap}{A rap object.}

\item{rules_ids}{The ids of the rules to be applied.}

\item{return_between}{Do you want to return the middle states?}
}
\value{
A new rap object, the result of applying the given rules.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Apply one or more ruless to a given P system given as a \code{rap} object.
}
\section{TODO}{

\itemize{
\item Include examples
}
}

\seealso{
\code{apply_rule_menv} to apply one rule in a multienvironmental P system, \code{apply_rule} to apply one rule in a monoenvironmental P system.
}
