% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_affected_comps.R
\name{get_affected_comps}
\alias{get_affected_comps}
\title{Template for new functions}
\usage{
get_affected_comps(rules)
}
\arguments{
\item{rules}{A set of rules from a \code{rap} object.}
}
\value{
It returns...
}
\description{
This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{Warning}{

This is a warning
}

\examples{
print("TODO:")

}
