% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_propensities.R
\name{get_propensities}
\alias{get_propensities}
\title{Template for new functions}
\usage{
get_propensities(rap, verbose = 0, debug = FALSE)
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{Warning}{

This is a warning
}

\examples{
Some examples

}
