% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rule_by_id.R
\name{get_rule_by_id}
\alias{get_rule_by_id}
\title{Template for new functions}
\usage{
get_rule_by_id(rap, rule_id)
}
\arguments{
\item{rap}{A \code{rap} object.}

\item{rule_id}{The id of the rule we are looking for.}
}
\value{
It returns...
}
\description{
This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{Warning}{

This is a warning
}

\examples{
Some examples

}
