% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_applicable.R
\name{is_applicable}
\alias{is_applicable}
\title{Template for new functions}
\usage{
is_applicable(rap, rule_info, verbose = 0)
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{Warning}{

This is a warning
}

\examples{
Some examples

}
