% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_demo_dataset.R
\name{load_demo_dataset}
\alias{load_demo_dataset}
\title{Template for new functions}
\usage{
load_demo_dataset(dataset = NULL)
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
This is a template.
}
\section{Warning}{

This is a warning
}

\examples{
Some examples
}
