% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path2rap.R
\name{path2rap}
\alias{path2rap}
\title{Load a P System given as a path/URL to a XML/JSON file}
\usage{
path2rap(
  path,
  use_codification = FALSE,
  verbose = 5,
  demo = FALSE,
  debug = FALSE
)
}
\arguments{
\item{path}{Path/URL to the input file, given as a \code{.xml}.}

\item{verbose}{Level of verbosity, between 0 and 5.}

\item{demo}{To execute as a demo.}

\item{debug}{To run in debugging mode, useful in order to catch errors.}
}
\value{
A rap object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This is the basic function of the package, and allows us to read a given P system, turning it into a rap object.
}
\section{Future work}{

\itemize{
\item Include different formats for inputs, like JSON or even the \code{.pli} itself.
}
}

\examples{
path = "https://raw.githubusercontent.com/Xopre/psystems-examples/main/plingua-5.0/RAPS/BIG/FAS.xml"
my_rap = path2rap(path)
show_rap(my_rap)

}
