% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_multiset.R
\name{print_multiset}
\alias{print_multiset}
\title{Improved representation of a processed multiset}
\usage{
print_multiset(multiset)
}
\arguments{
\item{multiset}{The processed multiset.}
}
\value{
Nothing, it only prints.
}
\description{
This function has not been tested enough yet.
}
\section{Warning}{

Experimental function.
}

\examples{
TODO
example_p_system = read_xml_p_system()
read_xml_p_system(demo = FALSE, path = "https://raw.githubusercontent.com/Xopre/psystems-examples/main/plingua/transition_full.xml")
}
