% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_multiset.R
\name{process_multiset}
\alias{process_multiset}
\title{Process a given multiset given as a XML nodeset}
\usage{
process_multiset(xml_multiset, label, n_membranes)
}
\arguments{
\item{xml_multiset}{The xml2 variable with the objects.}

\item{label}{A vector of labels for the different membranes.}

\item{n_membranes}{The number of membranes in the considered system.}
}
\value{
A processed multiset, given as a tibble.
}
\description{
This function has not been tested enough yet.
}
\section{Warning}{

Experimental function.
}

\examples{
TODO
example_p_system = read_xml_p_system()
read_xml_p_system(demo = FALSE, path = "https://raw.githubusercontent.com/Xopre/psystems-examples/main/plingua/transition_full.xml")
}
