% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_p_system_5.R
\name{read_p_system_5}
\alias{read_p_system_5}
\title{Load a P System given as a XML/JSON file}
\usage{
read_p_system_5(demo_mode = TRUE, path = NULL, verbose = 5)
}
\arguments{
\item{demo_mode}{Use demo settings? If FALSE path is mandatory.}

\item{path}{Path to the input file if necessary.}

\item{verbose}{Level of verbosity, between 0 and 5.}
}
\value{
A list with a loaded tibble and the properties of the read system.
}
\description{
This functions admits a demo mode, loading predefined systems. It is the basic function of the package.
}
\section{Warning}{

Experimental function.
}

\examples{
TODO
}
