% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xml_p_system.R
\name{read_xml_p_system}
\alias{read_xml_p_system}
\title{Load a P System given as a XML file}
\usage{
read_xml_p_system(path, verbose = TRUE, print = TRUE)
}
\arguments{
\item{path}{Path to the input file}

\item{verbose}{Verbose?}

\item{print}{Show read data?}
}
\value{
A loaded data.frame
}
\description{
Bla bla bla
}
