% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_rap.R
\name{show_rap}
\alias{show_rap}
\title{Template for new functions}
\usage{
show_rap(my_param)
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
This is a template.
}
\section{Warning}{

This is a warning
}

\examples{
Some examples
}
