% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_rap_objects_evolution.R
\name{show_rap_objects_evolution}
\alias{show_rap_objects_evolution}
\title{Print info of a given \code{rap} object}
\usage{
show_rap_objects_evolution(rap_objects)
}
\arguments{
\item{rap}{The \code{rap} object.}
}
\value{
Nothing, only prints.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Print objects of a given \code{rap} object.
}
\examples{
rap = RAPS::path2rap(demo = 1)
focus_on = list("MEM" = 2:3, "OBJ")
show_rap(rap, focus_on)
focus_on = list("RUL")

}
