% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate the evolution of a given P system}
\usage{
simulate(rap, n_steps)
}
\arguments{
\item{rap}{A rap object.}

\item{n_steps}{The number of evolutive steps to simulate.}
}
\value{
A rap object.
}
\description{
TODO: The description.
}
\section{Warning}{

This function comes with no warranty.
}

\examples{
TODO.
}
