% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gil.R
\name{simulate_gil}
\alias{simulate_gil}
\title{Template for new functions}
\usage{
simulate_gil(rap_environment)
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
This is a template.
}
\section{Warning}{

This is a warning
}

\examples{
Some examples
}
