% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_one.R
\name{simulate_one}
\alias{simulate_one}
\title{Simulate one evolutionary step in a given P system}
\usage{
simulate_one(rap)
}
\arguments{
\item{rap}{A rap object}
}
\value{
A rap object
}
\description{
TODO: The description.
}
\section{Warning}{

This function comes with no warranty.
}

\examples{
TODO: Some examples
}
