% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_p_system.R
\name{simulate_p_system}
\alias{simulate_p_system}
\title{Simulate the evolution of a loaded P System}
\usage{
simulate_p_system(
  p_system,
  verbose = TRUE,
  save = FALSE,
  semantic_dir = "semantics/transition_semantics_1.txt"
)
}
\arguments{
\item{p_system}{The p_system whose evolution we want to simulate.}

\item{verbose}{Verbose?}

\item{save}{Do you want to save the results? TODO}

\item{semantic_dir}{Directory in which the semantics of the system are written. TODO}
}
\value{
A list with a loaded data.frame and the properties of the read system
}
\description{
Given a P System and some semantics we want to follow, simulate the evolution of the system.
}
\section{Warning}{

Function under development.
DEPRECATED IN FAVOR OF \code{simulate}
}

\examples{
example_p_system = read_xml_p_system()
simulate_p_system(example_p_system)
}
\seealso{
\code{\link[=read_xml_p_system]{read_xml_p_system()}} for loading P Systems
}
