% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url2rap.R
\name{url2rap}
\alias{url2rap}
\title{Template for new functions}
\usage{
url2rap(my_param)
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
This is a template.
}
\section{Warning}{

This is a warning
}

\examples{
Some examples
}
