% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_det_menv.R
\name{alg_det_menv}
\alias{alg_det_menv}
\title{Template for new functions}
\usage{
alg_det_menv(rap, max_T = 10, debug = FALSE)
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
This is a template.
}
\section{Warning}{

This is a warning
}

\examples{
Some examples
}
