% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_det_menv.R
\name{alg_det_menv}
\alias{alg_det_menv}
\title{Deterministic waiting time algorithm for multienvironmental systems}
\usage{
alg_det_menv(
  rap,
  max_T = 1e-05,
  verbose = 2,
  debug = FALSE,
  debug_trinity = FALSE,
  random_trinity_selection = FALSE,
  save_each = NULL
)
}
\arguments{
\item{rap}{A \code{rap} object, usually generated with \code{load_demo_dataset()} or \code{path2rap()}.}

\item{max_T}{Maximum simulation time.}

\item{verbose}{The verbosity, between 0 and 5.}

\item{debug}{Useful if you want to debug an execution of the function.}

\item{debug_trinity}{The same as \code{debug} but applied only to the debugging of the generation of the trinities.}

\item{random_trinity_selection}{If there are more than one possible trinity, do you want to select one randomly (TRUE) or do you want to select the first one according to \code{dplyr::arrange()} (FALSE)?}

\item{save_each}{A \verb{function(rap)}. If not \code{NULL}, a list of \verb{function(rap)} is returned.}
}
\value{
A new \code{rap} object, the result of simulating the multienvironmental deterministic algorithm.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Simulates the evolution of a P system given as a \code{rap} object using the deterministic waiting time algorithm.
}
\section{TODO}{

\itemize{
\item Add references.
\item Improve efficiency saving the trinities.
}
}

\examples{
fas_path = "https://raw.githubusercontent.com/Xopre/psystems-examples/main/plingua-5.0/RAPS/BIG/FAS.xml"
# my_rap = path2rap(fas_path)
# new_rap = alg_det_menv(my_rap, max_T = 1e-6, verbose = 4)
# show_rap(new_rap)

}
