% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_gillespie.R
\name{alg_gillespie}
\alias{alg_gillespie}
\title{Gillespie's algorithm for monoenvironmental systems}
\usage{
alg_gillespie(
  rap,
  max_T = 10,
  propensity_function = RAPS::get_propensities,
  verbose = 2,
  debug = FALSE,
  debug_pair = FALSE,
  random_pair_selection = FALSE,
  save_each = NULL
)
}
\arguments{
\item{rap}{A \code{rap} object, usually generated with \code{load_demo_dataset()} or \code{path2rap()}.}

\item{max_T}{Maximum simulation time.}

\item{propensity_function}{A function to update propensities during each step}

\item{verbose}{The verbosity, between 0 and 5.}

\item{debug}{Useful if you want to debug an execution of the function.}

\item{debug_pair}{Used for debugging the creation of the pairs.}

\item{random_pair_selection}{If two pairs are available, do you want to choose one at random?}

\item{save_each}{A \verb{function(rap)}. If not \code{NULL}, a list of \verb{function(rap)} is returned.}
}
\value{
A new \code{rap} object, the result of simulating the Gillespie's algorithm for monoenvironmental systems.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Simulates the evolution of a P system given as a \code{rap} object using the Gillespie's algorithm for monoenvironmental systems
}
\section{TODO}{

Add references.
}

\examples{
fas_path = "https://raw.githubusercontent.com/Xopre/psystems-examples/main/plingua-5.0/RAPS/BIG/FAS.xml"
# my_rap = path2rap(fas_path)
# new_rap = alg_gillespie(my_rap, verbose = 4)
# show_rap(new_rap)

}
