% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_gillespie.R
\name{alg_gillespie}
\alias{alg_gillespie}
\title{Template for new functions}
\usage{
alg_gillespie(rap, max_T = 100, return_middle_states = TRUE)
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{Warning}{

This is a warning
}

\examples{
Some examples
}
