% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_gillespie_kernel.R
\name{alg_gillespie_kernel}
\alias{alg_gillespie_kernel}
\title{Kernel of Gillespie's algorithm}
\usage{
alg_gillespie_kernel(rules)
}
\arguments{
\item{rules}{A set of rules}
}
\value{
A rule to be executed, given as a \code{tibble} object with fields \code{j_c}, the rule id, and \code{tau_c}, the execution time of the rule.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Kernel of Gillespie's algorithm, required for both multienvironmental and monoenvironmental simulations.
}
\section{TODO}{

Add references.
}

