% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg_gillepies_menv.R
\name{alg_gillespie_menv}
\alias{alg_gillespie_menv}
\title{Gillespie's algorithm for multienvironmental systems}
\usage{
alg_gillespie_menv(rap, max_T = 1e-05, verbose = 2, debug = FALSE)
}
\arguments{
\item{rap}{A \code{rap} object, usually generated with \code{load_demo_dataset()} or \code{path2rap()}.}

\item{max_T}{Maximum simulation time.}
}
\value{
A new \code{rap} object, the result of simulating the Gillespie's algorithm for multienvironmental systems.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Simulates the evolution of a P system given as a \code{rap} object using the Gillespie's algorithm for multienvironmental systems
}
\section{CAUTION}{

For now affected compartment is equal to every compartment, like in the deterministic waiting time algorithm.
}

\examples{
fas_path = "https://raw.githubusercontent.com/Xopre/psystems-examples/main/plingua5/RAPS/BIG/FAS.xml"
my_rap = path2rap(fas_path)
new_rap = alg_gillespie_menv(my_rap, verbose = 4)
show_rap(new_rap)

}
