% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rule.R
\name{apply_rule}
\alias{apply_rule}
\title{Apply a rule to a given rap object}
\usage{
apply_rule(rap, rule_id, verbose = FALSE, debug = FALSE)
}
\arguments{
\item{rap}{A rap object.}

\item{rule_id}{The id of the rule to be applied.}

\item{...}{Other parameters such as \code{debug} or \code{verbose}.}
}
\value{
A new rap object, the result of applying the given rule.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
TODO
}
\seealso{
\code{apply_rules} to apply more than one rule at the same time.
#' @section TODO:
\itemize{
\item Print the trace of the execution, perhaps with the \code{RAPS::show_rap()} function.
}
}
