% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rule.R
\name{apply_rule}
\alias{apply_rule}
\title{Apply a rule to a given rap object}
\usage{
apply_rule(rap, rule_id)
}
\arguments{
\item{rap}{A rap object-}

\item{rule_id}{The id of the rule to be applied-}
}
\value{
A new rap object, the result of applying the given rule-
}
\description{
Apply a rule to a given rap object
}
\examples{
TODO
}
\seealso{
\code{apply_rules} to apply more than one rule at the same time.
}
