% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_applicability.R
\name{check_applicability}
\alias{check_applicability}
\title{Template for new functions}
\usage{
check_applicability(
  verbose,
  affected_membranes,
  main_membrane_index,
  rule_info
)
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
This is a template.
}
\section{Warning}{

This is a warning
}

\examples{
Some examples
}
