% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_rule.R
\name{choose_rule}
\alias{choose_rule}
\title{Choose a rule to be applied in a rap object}
\usage{
choose_rule(rap, semantics = "GIL")
}
\arguments{
\item{rap}{A rap object.}

\item{semantics}{The path to a semantics object, or the id of a given algorithm from the following list: GIL (for Gillespie), DET (for Deterministic Waiting Times Algorithm).}
}
\value{
The id of a rule which is to be applied.
}
\description{
Note that the rule is not applied.
}
\examples{
TODO.
}
