% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_concentration_of.R
\name{get_concentration_of}
\alias{get_concentration_of}
\title{Template for new functions}
\usage{
get_concentration_of(rap, membrane_id, chosen_object, is_coded = FALSE)
}
\arguments{
\item{rap}{A \code{rap} object.}

\item{membrane_id}{The id of the membrane in which we are looking for \code{chosen_object}.}

\item{chosen_object}{The id or the coded id of the object we are looking for.}

\item{is_coded}{Is \code{chosen_object} coded?}
}
\value{
It returns...
}
\description{
This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{Warning}{

This is a warning
}

\examples{
Some examples

}
