% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_propensities.R
\name{get_propensities}
\alias{get_propensities}
\title{Template for new functions}
\usage{
get_propensities(rap, verbose = 0, debug = FALSE)
}
\arguments{
\item{rap}{A \code{rap} object.}

\item{verbose}{The verbosity, between 0 and 5.}

\item{debug}{Useful if you want to debug an execution of the function.}
}
\value{
It returns...
}
\description{
This is a template.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{Warning}{

This is a warning
}

\examples{
print("TODO:")

}
