% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_omit.R
\name{na_omit}
\alias{na_omit}
\title{Omit all NAs in a given vector}
\usage{
na_omit(v)
}
\arguments{
\item{v}{The vector.}
}
\value{
The vector v without the NAs.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Used as help in some functions
}
\section{Warning}{

Experimental function.
}

\examples{
my_vector = c(1, 2, NA, 4, NA)
na_omit(my_vector)
}
