% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path2rap.R
\name{path2rap}
\alias{path2rap}
\title{Load a P System given as a path/URL to a XML/JSON file}
\usage{
path2rap(path = NULL, verbose = 5, demo = 1, max_depth = 3)
}
\arguments{
\item{path}{Path/URL to the input file.}

\item{verbose}{Level of verbosity, between 0 and 5.}

\item{demo}{To execute as a demo.}

\item{max_depth}{To look for new structures in the jerarchy of the P system.}
}
\value{
A rap object.
}
\description{
This is the basic function of the package, and allows us to read a given P system, turning it into a rap object.
}
\section{Warning}{

Experimental function.
}

\section{Future work}{

\itemize{
\item Include different formats for inputs, like JSON or even the \code{.pli} itself.
}
}

\examples{
TODO.
}
