% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAPS-package.R
\docType{package}
\name{RAPS-package}
\alias{RAPS}
\alias{RAPS-package}
\title{RAPS: 🎤 RAPS: R Aid for P Systems}
\description{
In this paragraph I will explain what the package does. The description continues here.
}
\author{
\strong{Maintainer}: Xopre - \email{rodgal2020@gmail.com} (\href{https://orcid.org/0000-0002-5825-2724}{ORCID})

}
\keyword{internal}
