% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xml_p_system.R
\name{read_xml_p_system}
\alias{read_xml_p_system}
\title{Load a P System given as a XML file}
\usage{
read_xml_p_system(demo_mode = TRUE, path = NULL, verbose = TRUE)
}
\arguments{
\item{demo_mode}{Use demo settings? If not path is mandatory.}

\item{path}{Path to the input file if necessary.}

\item{verbose}{Verbose?}
}
\value{
A list with a loaded data.frame and the properties of the read system
}
\description{
Bla bla bla
}
