% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xml_p_system.R
\name{read_xml_p_system}
\alias{read_xml_p_system}
\title{Load a P System given as a XML file}
\usage{
read_xml_p_system(demo_mode = TRUE, path = NULL, verbose = TRUE)
}
\arguments{
\item{demo_mode}{Use demo settings? If FALSE path is mandatory.}

\item{path}{Path to the input file if necessary.}

\item{verbose}{Verbose?}
}
\value{
A list with a loaded data.frame and the properties of the read system
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\strong{Deprecation info}: Deprecated in favor of \code{path2rap}.
This functions admits a demo mode, loading predefined systems. It is the basic function of the package.
}
\section{Warning}{

Experimental function.
}

\examples{
example_p_system = read_xml_p_system()
read_xml_p_system(demo = FALSE, path = "https://raw.githubusercontent.com/Xopre/psystems-examples/main/plingua/transition_full.xml")
}
