% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_rule.R
\name{show_rule}
\alias{show_rule}
\title{Print info of a given rule from a \verb{rap object}}
\usage{
show_rule(rule)
}
\arguments{
\item{rule}{The rule from a \code{rap} object. \strong{NOT} the id.}
}
\value{
Nothing, only prints.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Print a rule from a \code{rap} object.
}
\section{TODO}{

\itemize{
\item Improve visualization.
\item Add tibble-less mode in order to avoid unnecessary extra tibble info.
}
}

\examples{
# rap = RAPS::path2rap(demo = 1)
# show_rule(rap$Rules[1, ])

}
