% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHECKME_simulate_gil.R
\name{simulate_gil}
\alias{simulate_gil}
\title{Template for new functions}
\usage{
simulate_gil(rap_environment, max_T = 100)
}
\arguments{
\item{my_param}{Yep, it's a parameter.}
}
\value{
It returns...
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This is a template.
}
\section{Warning}{

This is a warning
}

\examples{
Some examples
}
